move otlp export and helpers to otlp file
